hf_space_deploy/server/requirements.txt
fastapi>=0.104.0
pydantic>=2.0.0
uvicorn>=0.24.0
websockets>=15.0.0
httpx>=0.28.0
gradio>=4.0.0
openenv-core>=0.2.3
